pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ccspr"
version = "0.1.0"
description = "Reproducible CC-SPR research pipeline"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "scipy>=1.10",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "networkx>=3.0",
  "gudhi>=3.8",
  "cvxpy>=1.4",
  "GraphRicciCurvature>=0.5.3",
  "PyYAML>=6.0",
  "requests>=2.31",
  "joblib>=1.3",
  "anndata>=0.10",
  "scanpy>=1.10",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
